CLI class
