+++ release_bot/releasebot.py
@@ -5,41 +5,14 @@ into various downstream services
 import re
 import time
 import os
-import argparse
-import logging
 from semantic_version import Version, validate
 from sys import exit
 
+from .cli import CLI
 from .configuration import configuration
 from .fedora import Fedora
 from .github import Github
 from .pypi import PyPi
-
-
-def parse_arguments():
-    """Parse application arguments"""
-    parser = argparse.ArgumentParser(description="Automatic releases bot", prog='release-bot')
-    parser.add_argument("-d", "--debug", help="turn on debugging output",
-                        action="store_true", default=False)
-    parser.add_argument("-c", "--configuration", help="use custom YAML configuration",
-                        default='')
-    parser.add_argument("-v", "--version", help="display program version", action='version',
-                        version=f"%(prog)s {configuration.version}")
-    parser.add_argument("-k", "--keytab", help="keytab file for fedora", default='')
-
-    args = parser.parse_args()
-    if args.configuration:
-        path = args.configuration
-        if not os.path.isabs(path):
-            args.configuration = os.path.join(os.getcwd(), path)
-        if not os.path.isfile(path):
-            configuration.logger.error(
-                f"Supplied configuration file is not found: {args.configuration}")
-            exit(1)
-    if args.debug:
-        configuration.logger.setLevel(logging.DEBUG)
-    for key, value in vars(args).items():
-        setattr(configuration, key, value)
 
 
 class ReleaseBot:
@@ -133,7 +106,7 @@ class ReleaseBot:
 
 
 def main():
-    parse_arguments()
+    CLI.parse_arguments()
     configuration.load_configuration()
 
     rb = ReleaseBot(configuration)
